requirements.txt
aiofiles==22.1.0
anyio==3.6.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
decorator==4.4.2
dnspython==2.2.1
fastapi==0.85.0
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.22.1
imageio-ffmpeg==0.4.7
moviepy==1.0.3
proglog==0.1.10
pycparser==2.21
pymongo==4.2.0
python-dotenv==0.21.0
python-magic==0.4.27
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
tqdm==4.64.1
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
